Combine frontend and backend into traditional monorepo
This commit creates one node project that can run either the frontend, backend, or both via one set of node scripts. Configuration files are moved to the top-level of the project. In order to migrate your branches, make sure to copy your private file (.env and the firebase SDK configuration) to the top-level of the project, as well as delete the node_modules folder from the frontend and backend folders (and running 'npm i' from the main level after rebasing).